Extracts Webmock setup into a separate class
